added Auth0... need to re-wire from App.js to App_old.js
