+++ src/auth_config.json
@@ -1,0 +1,4 @@
+{
+    "domain": "dev-9zf3-xc1.auth0.com",
+    "clientId": "gAanIirzJ76DQa5mSkSlp1lGyn8UkdGs"
+  }
