+++ src/components/Profile.js
@@ -1,0 +1,27 @@
+import React, { Fragment } from "react";
+import { useAuth0 } from "../react-auth0-spa";
+
+const Profile = () => {
+  const { loading, user } = useAuth0();
+
+  if (loading || !user) {
+    return <div>Loading...</div>;
+  }
+
+console.log(user)
+
+  return user ? (
+    <Fragment>
+      <img src={user.picture} alt="Profile" />
+
+      <h2>{user.name}</h2>
+      <p>{user.email}</p>
+      <code>{JSON.stringify(user, null, 2)}</code>
+
+    </Fragment>
+  ) :
+ <div>(user)</div> 
+  ;
+};
+
+export default Profile;
